src/requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.3.6
channels==4.1.0
channels-redis==4.2.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.0.4
django-cors-headers==4.3.1
django-extensions==3.2.3
django-redis==5.4.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
gunicorn==22.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
Hypercorn==0.17.3
hyperframe==6.0.1
inflection==0.5.1
kombu==5.3.7
msgpack==1.0.8
packaging==24.0
pillow==10.3.0
priority==2.0.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.4
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
wsproto==1.2.0
